Update bun to 1.2.2 and fix lints
