+++ frontend/src/ai/OpenAIContextDef.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import OpenAI from "openai";
+
+export const OpenAIContext = createContext<OpenAI | undefined>(undefined);
