+++ frontend/src/components/Explainer.tsx
@@ -38,7 +38,7 @@ function VerificationStatus() {
       }
     }
     verify();
-  }, []);
+  }, [os]);
 
   return (
     <Link to="/proof" className="flex items-center gap-1 text-sm font-medium hover:underline">
